requirements.txt
celery==5.6.2
fastapi==0.129.2
httpx==0.28.1
psycopg2-binary==2.9.11
pydantic==2.12.5
PyJWT==2.11.0
redis==7.2.0
SQLAlchemy==2.0.46
uvicorn==0.41.0
